requirements.txt
rdkit-pypi==2022.3.5
scikit-learn==0.23.2
scipy==1.7.3
jsonify==0.5
jupyterlab==3.4.5
pandas==1.3.5
joblib==1.1.0
matplotlib==3.5.3
shap==0.41.0
catboost==1.0.6
# sklranger==0.7.0
tensorboard==1.13.1
tensorflow==1.13.2
#xsmiles==0.1.3 #TODO uncomment when available through pip and update README